Switch all validators to use a parameter object pack.
This is easier for users to gradually assemble parameters, rather than having
to remember a long function signature for each validator.

To avoid copying large objects, we introduce a WrappedOption wrapper that
accepts a const pointer to an externally owned object, or just uses an
internally owned instance of that object. This allows users to use a const
reference if they don't want to the parameter pack to take ownership.