Optionally return parsed CSV contents after validation. (#8)
This allows callers to validate the CSV and return the contents, avoiding the
need for multiple passes through the CSV if the CSV's contents are needed for
validation of another file. An archetypal example is the need to load the
seqnames from the sequence_information to validate the genomic_ranges; this can
now be done systematically for all CSV-based formats to avoid the performance 
penalty of non-random-access during validation.

Note that this is less of an issue for HDF5, as (i) HDF5 supports random access
so we can just pull out the fields of interest afterwards and (ii) most HDF5
specs are fairly self-contained and don't need inter-file validation.