Improved the HDF5 data frame format to make it more self-contained. (#9)
This moves the type/format/ordered specifications inside the HDF5 file, so that
readers can do a reasonable job reconstructing the DF without any input from
the metadata. Factor levels are also moved into the file, which is the biggest
change as the factors are now groups instead of datasets.

This change is handled by a new 'version' attribute on the parent group, which
determines whether to use the new format (if the attribute is present) or to
use the old specification (the specifics of which are determined from the
various 'version'-related properties in the hdf5_data_frame schema).

The general philosophy is for the metadata to be a mirror of the information in
the file; some redundancy is accepted in return for ease of use.