Improved the sparse matrix spec to be more self-contained. (#10)
This adds version and format attributes to the main group so that it's possible
to parse it without any reference to the schema. We also move the names into
a well-defined subgroup of the main group to eliminate the need to check the
name of the dimname-containing group in the metadata properties.