Overhauled the dense array spec to be more self-contained. (#12)
This adds version and type attributes to the main dataset so that it's possible
to parse it without any reference to the schema. We also add a dimnames
attribute, which really is a poor man's dimension scales. (We don't use H5DS
itself because it's outside the core HDF5 lib and I don't want to link to it).

As part of this change, we consolidated a lot of checks that were common to
both the HDF5 sparse matrix 'data' and the (new) HDF5 dense array dataset. We
also consolidated the tests for the new dimnames attribute. This reduces the
amount of repetition and improves consistency between these two validators.