Make the 'dimension-names' attribute optional.
Simplifies the validators by eliminating the need to specify whether an
array/matrix has dimension names prior to running the validators.

It also avoids us having to carry around that information in the schema, given
that the existence of dimension names doesn't really add much to the decision
making process about an array/matrix in most cases.