Begin transitioning to a simpler layout for the on-disk representation. (#13)
This eliminates the need to consult the schema and/or a dedicated metadata
file. Now, we just have an OBJECT file that specifies the object type, and that
allows us to dispatch to all the various validators for each type.

Each representation now uses hard-coded file and directory names, which avoids
the need to check user-defined file names in the metadata. Also, because each
object's representation is fully defined by its directory contents, we can
greatly simplify the takane API by avoiding the need for loads of arguments.

For the time being, I've only simplified the basic object classes (defined in
S4Vectors); the other classes will eventually transition to this new scheme.