Formally specify which objects implement each interface.
This allows classes to be composed of children of variable types provided that
the various possible types satisfy a interface for the object. For example, we
can have different varieties of data frames with custom on-disk formats, and
any of them could be used as the mcols of an S4Vectors::Vetor object.