SCEs now have a dedicated JSON file to store version, main exp name.
This is particularly necessary to hold the version string, so as to
future-proof the layout in case of version-specific behaviors; then
clients at least have the ability to find out what version it is.

We also require the main experiment name, if supplied, to be unique from
the alternative experiments, as otherwise it wouldn't be swappable.