Convert OBJECT back to a JSON-formatted metadata file to store the version.  (#20)
All versions should now stored in the top-level metadata rather than in the
HDF5 files. This is motivated by the philosophy that, for multi-file
representations, it doesn't make sense to look inside the HDF5 file for the
version of the entire object; there's nothing particularly special about the
HDF5 file, and there might be multiple (or no) HDF5 files for a single object.

By moving it to the top-level OBJECT, readers and validators can more clearly
see the version of the object's representation, and coordinate the treatment of
multiple files accordingly. It also resolves a clash between the version of the
simple_list and the uzuki version, which are subtly different as the latter does
not consider the organization of the external object subdirectory.

A consequence of this change is that the top-level OBJECT now needs to store
more than just the type, and thus returns to being a JSON-formatted file with
at least a 'type' property. Each object may also store its own metadata inside
a property named after the object class, allowing consolidation of the various
SE metadata into a single file for simplicity. Versions are expected to be
stored inside the custom object properties, which enables variable versioning
of the base and derived classes.