More stringent requirements on the FASTA/Q file metadata.
Both now must report the sequence type in their metadata, and the FASTQ files
must report the quality type and offsets explicitly. This is necessary to
interpret the content of the files, so we might as well add it; and it allows
us to be consistent with the metadata properties in the sequence_string_set.