Implement validator for VCF experiments. (#28)
These are VCF files that can be interpreted as summarized_experiment objects,
i.e., the number of samples is used as the number of columns, and the number of
rows is the number of records (or alternative alleles, if expanded). 

We store the dimensions in the OBJECT file to avoid having to do a whole file
scan in height() and dimensions(). That said, we still do a whole-file scan in
validate() to check that the number of samples and records are correct.  We
also check the file format string of the VCF file, but it is not otherwise
subject to additional validation because that is tedious.