Prohibit non-positive starts for non-circular reference sequences.
This is slightly more strict than the previous behavior, where non-positive
starts were allowed for non-circular chromosomes that had missing sequence
lengths. This change is motivated by the idea that the sequence length has no
bearing on the illegality of a non-positive start for non-circular references.

In practice, this probably won't change much as it only triggers if the
sequence is known to be non-circular, and most of the time, the circular flag
is just missing because most users don't fill out the Seqinfo on GRanges.