Distinguish between interface satisfication and derived types.
The latter is a stricter version of the former, as the validation code can
assume that specific files/formats are present in derived types. This is not
possible if the interface is just satisfied, as that describes a user
experience and doesn't provide any guarantees with respect to the files.

Also modified all code that was previously checking for hard-coded types to use
the new derived_from() function, to allow future usage of derived types.