Avoid GCC's new dangling reference warnings from string temporaries. (#42)
These are all false positive warnings but GCC seems unlikely to change. So, we
just create std::strings from string literals outside of the function call, and then
pass them in; this should shut up GCC without silencing the warning outright.

Incidentally, the availability of separate variable allows us to remove some
duplicate instances of the hard-coded strings, so that's a nice side-effect.