Introduce support for custom variable length strings in some object types. (#44)
This uses a custom layout to overcome the weakness of HDF5 for efficiently
storing arrays of variable length strings. For now, we only define it for
atomic_vector, dense_array and data_frame columns; this has the most impact but
avoids spreading it too much as it might be made obsolete by HDF5v2.